1-basic/pyproject.toml
[project]
name = "1-basic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "logfire>=3.14.0",
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
]
